refactor: break schema out separately
